GUI/requirements.txt
PySide6==6.1.1
paho-mqtt==1.5.1